+++ src/com/markehme/factionsalias/entities/FactionsCommandSkeleton.java
@@ -1,15 +1,16 @@
-package com.markehme.factionsalias.support;
+package com.markehme.factionsalias.entities;
 
 import java.util.List;
 
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
-import org.bukkit.command.CommandSender;
-
+import org.bukkit.entity.Player;
 
 import me.markeh.factionsframework.command.FactionsCommand;
+import me.markeh.factionsframework.command.requirements.ReqHasFaction;
+import me.markeh.factionsframework.command.requirements.ReqIsPlayer;
 
-public class FactionsPlusXCommandSkeleton extends FactionsCommand {
+public class FactionsCommandSkeleton extends FactionsCommand {
 	public String exec = "";
 	public String permissionRequired = null;
 	public String permissionDeniedMsg = "";
@@ -20,7 +21,7 @@ public class FactionsPlusXCommandSkeleton extends FactionsCommand {
 	
 	public boolean requiresIsPlayer = false;
 	
-	public FactionsPlusXCommandSkeleton(
+	public FactionsCommandSkeleton(
 			List<String> aliases,
 			boolean requiresFactionsEnabled,
 			boolean requiresIsPlayer,
@@ -37,13 +38,10 @@ public class FactionsPlusXCommandSkeleton extends FactionsCommand {
 			this.aliases.add((String) alias);
 		}
 		
-		if(requiresIsPlayer) {
-			this.requiresIsPlayer = true;
-		}
+		if(requiresIsPlayer) this.addRequirement(ReqIsPlayer.get());
 		
-		if(requiresInFaction) {
-			this.mustHaveFaction = true;
-		}
+		if(requiresInFaction) this.addRequirement(ReqHasFaction.get());
+		
 		
 		if(requiresFactionsEnabled) {
 			this.requiresFactionsEnabled = true;
@@ -56,6 +54,7 @@ public class FactionsPlusXCommandSkeleton extends FactionsCommand {
 		this.permissionRequired = permission;
 		this.permissionDeniedMsg = permission;
 		this.description = desc;
+		this.helpLine = desc;
 		
 		exec = executingCommand;
 		
@@ -66,28 +65,27 @@ public class FactionsPlusXCommandSkeleton extends FactionsCommand {
 
 	@Override
 	public void run() {
-		if (this.requiresIsPlayer && player == null) {
+		if (this.requiresIsPlayer && ! (sender instanceof Player)) {
 			msg("This is a player-only command.");
 			return;
 		}
 		
-		if(permissionRequired != null && player != null) {
-			if( ! player.hasPermission(permissionRequired)) {
+		if (permissionRequired != null) {
+			if ( ! sender.hasPermission(permissionRequired)) {
 				msg(permissionDeniedMsg);
 				return;
 			}
 		}
 				
-		if(requirePlayerIsLeader && !fplayer.isLeader()) {
+		if (requirePlayerIsLeader && ! fplayer.isLeader()) {
 			msg(ChatColor.RED + "Only the leader of the faction can run this command.");
 			return;
 		}
 		
-		String argString = "";
-		for(String arg : this.arguments) {
-			argString += arg + " ";
-		}
-		Bukkit.getServer().dispatchCommand((CommandSender) player, exec + " " + argString.replaceAll("(&([a-f0-9]))", "& $2"));
+		String args = "";
+		if (this.arguments.size() > 0) args = this.getArgsConcated(0);
+		
+		Bukkit.getServer().dispatchCommand(sender, exec + " " + args.replaceAll("(&([a-f0-9]))", "& $2"));
 		
 	}
 }