+++ src/com/markehme/factionsalias/libs/gson/annotations/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package provides annotations that can be used with {@link com.markehme.factionsalias.libs.gson.Gson}.
+ * 
+ * @author Inderjeet Singh, Joel Leitch
+ */
+package com.markehme.factionsalias.libs.gson.annotations;