+++ src/com/markehme/factionsalias/libs/gson/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * This package provides the {@link com.markehme.factionsalias.libs.gson.Gson} class to convert Json to Java and
+ * vice-versa.
+ *
+ * <p>The primary class to use is {@link com.markehme.factionsalias.libs.gson.Gson} which can be constructed with
+ * {@code new Gson()} (using default settings) or by using {@link com.markehme.factionsalias.libs.gson.GsonBuilder}
+ * (to configure various options such as using versioning and so on).</p>
+ *
+ * @author Inderjeet Singh, Joel Leitch
+ */
+package com.markehme.factionsalias.libs.gson;